requirements.txt
# CICLOPS Backend - Dependencies
# PostgreSQL + OpenAI para análisis financiero

# Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database - PostgreSQL
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# AI - OpenAI + Anthropic
openai==1.12.0
anthropic==0.39.0

# PDF Processing
pdfplumber==0.10.3

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2

# Utils
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
